requirements.txt
libpebble2>=0.0.26
pydbus>=0.6.0
